requirements.txt
flask==2.3.3
qqwry-py3==1.2.1
flask-cors==4.0.0